bixby-medi/requirements.txt
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.6
django-extensions==2.2.5
django-heroku==0.3.1
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.10.3
gunicorn==19.9.0
haversine==2.1.2
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2==2.8.4
psycopg2-binary==2.8.4
pytz==2019.3
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
whitenoise==4.1.4
